pyproject.toml
[project]
name = "photogen-api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "pydantic-settings (>=2.8.1,<3.0.0)", "gunicorn (>=23.0.0,<24.0.0)", "tortoise-orm[asyncpg] (>=0.25.0,<0.26.0)", "python-jose[cryptography] (>=3.4.0,<4.0.0)", "aiogram (>=3.20.0,<4.0.0)", "python-multipart (>=0.0.20,<0.0.21)"]

[[project.authors]]
name = "Elisey"
email = "eliseyblajen@gmail.com"

[[tool.poetry.packages]]
include = "photogen_api"
from = "src"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.1"
tomlkit = "^0.13.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies.aerich]
extras = ["standard"]
version = "^0.8.2"

[tool.aerich]
tortoise_orm = "photogen_api.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
